Server/requirements.txt
paho-mqtt==1.6.1
numpy==1.24.3
scipy==1.10.1
websockets==11.0.3
asyncio
